fixed paths references to modules that became faulty due to the picker move
